split on int() type
The variable 'pool_size' is converted to type 'int' in all cases before the second if/elif/elif test so it is always a mistake to use a string operator (split, via '[:1]') at that point.  This was causing all attempts at installation with ZFS to fail, with this change I've been able to install Antergos with a ZFS filesystem for the root partition.